% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{block_indices.hatsa_projector}
\alias{block_indices.hatsa_projector}
\title{Extract block indices from a hatsa_projector object}
\usage{
block_indices.hatsa_projector(x, ...)
}
\arguments{
\item{x}{A \code{hatsa_projector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list of block indices.
}
\description{
These indices map rows of the scores matrix (s) to their respective blocks (subjects).
}
