% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{coef.hatsa_projector}
\alias{coef.hatsa_projector}
\title{Extract coefficients (mean aligned sketch) from a hatsa_projector object}
\usage{
\method{coef}{hatsa_projector}(object, ...)
}
\arguments{
\item{object}{A \code{hatsa_projector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The mean aligned sketch matrix (v).
}
\description{
Extract coefficients (mean aligned sketch) from a hatsa_projector object
}
