% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_helpers.R
\name{compute_gev_spectrum_diagnostics}
\alias{compute_gev_spectrum_diagnostics}
\title{Compute GEV Spectrum Diagnostics}
\usage{
compute_gev_spectrum_diagnostics(Lambda_GEV, lambda_max_thresh)
}
\arguments{
\item{Lambda_GEV}{A numeric vector of eigenvalues obtained from GEV.}

\item{lambda_max_thresh}{A numeric threshold used to categorize eigenvalues.
Eigenvalues with absolute value less than this are considered 'retained' or 'stable'.}
}
\value{
A list containing the following diagnostic statistics:
  \\itemize{
    \\item{n_eigenvalues: Total number of eigenvalues.}
    \\item{min_eigenvalue: Minimum eigenvalue.}
    \\item{max_eigenvalue: Maximum eigenvalue.}
    \\item{mean_eigenvalue: Mean of eigenvalues.}
    \\item{median_eigenvalue: Median of eigenvalues.}
    \\item{sd_eigenvalue: Standard deviation of eigenvalues.}
    \\item{n_below_thresh: Number of eigenvalues with absolute value < lambda_max_thresh.}
    \\item{prop_below_thresh: Proportion of eigenvalues with absolute value < lambda_max_thresh.}
    \\item{n_above_thresh: Number of eigenvalues with absolute value >= lambda_max_thresh.}
    \\item{prop_above_thresh: Proportion of eigenvalues with absolute value >= lambda_max_thresh.}
  }
}
\description{
Computes various diagnostic statistics for a vector of Generalized Eigenvalues (GEV).
}
\examples{
  Lambda_GEV_sample <- c(0.1, 0.5, 0.85, 0.95, 1.2)
  compute_gev_spectrum_diagnostics(Lambda_GEV_sample, lambda_max_thresh = 0.8)
}
