% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_simple.R
\name{configure_task_params}
\alias{configure_task_params}
\title{Configure Task Parameters}
\usage{
configure_task_params(method, base_config, ...)
}
\description{
Configure Task Parameters
}
\keyword{internal}
