% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_simple.R, R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-safe or operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x, }{y Values to compare.}
}
\description{
Returns \code{x} if it is not \code{NULL}, otherwise \code{y}.
}
\keyword{internal}
