% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_simple.R
\name{hatsa_task_preset}
\alias{hatsa_task_preset}
\title{Task-HATSA Configuration Presets}
\usage{
hatsa_task_preset(preset_name, method)
}
\description{
Task-HATSA Configuration Presets
}
\keyword{internal}
