% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_geometry.R
\name{matrix_logm_spd}
\alias{matrix_logm_spd}
\title{Matrix Logarithm of an SPD Matrix}
\usage{
matrix_logm_spd(S, regularize_epsilon = 1e-06)
}
\arguments{
\item{S}{A numeric, symmetric matrix.}

\item{regularize_epsilon}{Epsilon for regularization. Default 1e-6.}
}
\value{
The matrix logarithm of S (a symmetric matrix).
}
\description{
Computes the principal matrix logarithm of a symmetric positive-definite (SPD)
matrix. The input matrix is first regularized to ensure positive definiteness.
}
\examples{
S1 <- matrix(c(2.3, -0.3, -0.3, 3.6), 2, 2)
logS1 <- matrix_logm_spd(S1)
print(logS1)
}
