% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_geometry.R
\name{matrix_sqrt_spd}
\alias{matrix_sqrt_spd}
\title{Matrix Square Root of an SPD Matrix}
\usage{
matrix_sqrt_spd(S, regularize_epsilon = 1e-06)
}
\arguments{
\item{S}{A numeric, symmetric matrix.}

\item{regularize_epsilon}{Epsilon for regularization. Default 1e-6.}
}
\value{
The matrix square root of S (an SPD matrix).
}
\description{
Computes the principal matrix square root of a symmetric positive-definite (SPD)
matrix. The input matrix is first regularized to ensure positive definiteness.
The matrix square root S_sqrt is such that S_sqrt %*% S_sqrt = S.
}
\examples{
S1 <- matrix(c(2.3, -0.3, -0.3, 3.6), 2, 2)
S1_sqrt <- matrix_sqrt_spd(S1)
# S1_reconstructed_from_sqrt <- S1_sqrt \%*\% S1_sqrt
# all.equal(S1, S1_reconstructed_from_sqrt) # Should be TRUE
}
