% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_stage}
\alias{message_stage}
\title{Helper for printing stage messages, optionally only in interactive sessions}
\usage{
message_stage(message_text, interactive_only = FALSE)
}
\arguments{
\item{message_text}{Message to print.}

\item{interactive_only}{Logical, if TRUE, message only if session is interactive.}
}
\description{
Helper for printing stage messages, optionally only in interactive sessions
}
\keyword{internal}
