% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_accessors.R
\name{plot_hatsa}
\alias{plot_hatsa}
\title{Plot HATSA Results}
\usage{
plot_hatsa(x, type = c("eigenvalues", "rotation_quality", "alignment"), ...)
}
\arguments{
\item{x}{A hatsa_projector object}

\item{type}{Plot type: "eigenvalues", "rotation_quality", or "alignment"}

\item{...}{Additional plotting parameters}
}
\description{
Creates diagnostic plots for HATSA alignment results.
}
