% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_qc_plots.R
\name{plot_k_stability_hatsa}
\alias{plot_k_stability_hatsa}
\title{HATSA Quality Control and Stability Plots}
\usage{
plot_k_stability_hatsa(
  projector_list_over_k,
  metrics_to_plot = c("Hk", "CV_eigen_cov"),
  Hk_dispersion_stat = "mean",
  spd_cov_coeffs_options = list(),
  Hk_options = list(),
  verbose = FALSE
)
}
\arguments{
\item{projector_list_over_k}{A list of \code{hatsa_projector} objects.
Each object should be the result of a HATSA run with a different 'k'.
The list can be named with k values, or k will be extracted from each object.}

\item{metrics_to_plot}{A character vector specifying which metrics to plot.
Possible values are "Hk" (Alignment Homogeneity based on Riemannian dispersion
of covariance of aligned coefficients) and/or "CV_eigen_cov" (Mean Coefficient
of Variation of eigenvalues of covariance of aligned coefficients).
Default is \code{c("Hk", "CV_eigen_cov")}.}

\item{Hk_dispersion_stat}{Character, either "mean" (default) or "median".
Specifies which statistic from \code{riemannian_dispersion_spd} to use for Hk.}

\item{spd_cov_coeffs_options}{A list of additional arguments to pass to
\code{hatsa::get_spd_representations(..., type = "cov_coeffs")}.
For example, \code{list(spd_regularize_epsilon = 1e-6)}.}

\item{Hk_options}{A list of additional arguments to pass to
\code{hatsa::riemannian_dispersion_spd(...)}.
For example, \code{list(tangent_metric = "logeuclidean", use_geometric_median = FALSE)}.
Note that `use_geometric_median` will be overridden by `Hk_dispersion_stat == "median"`.}

\item{verbose}{Logical, if TRUE, prints progress messages. Default is FALSE.}
}
\value{
A \code{ggplot} object containing the stability plots, or NULL if
  no valid metrics can be computed.
}
\description{
This file contains functions for generating QC plots, such as k-stability plots,
for HATSA results.
}
\examples{
# Conceptual example:
# Assume projector_k3, projector_k4, projector_k5 are hatsa_projector objects
# k_list <- list("3" = projector_k3, "4" = projector_k4, "5" = projector_k5)
# plot_k_stability_hatsa(k_list)

# To run if example data were available:
# if (requireNamespace("ggplot2", quietly = TRUE) &&
#     exists("generate_synthetic_hatsa_output_list")) {
#   example_projector_list <- generate_synthetic_hatsa_output_list(
#      k_values = c(3, 4, 5, 6), N_subjects = 5, V_p = 20
#   )
#   plot_k_stability_hatsa(example_projector_list)
# }
}
