% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{print.hatsa_projector}
\alias{print.hatsa_projector}
\title{Print method for hatsa_projector objects}
\usage{
\method{print}{hatsa_projector}(x, ...)
}
\arguments{
\item{x}{A \code{hatsa_projector} object.}

\item{...}{Additional arguments passed to print.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a concise summary of the hatsa_projector object.
}
