% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{print.summary.hatsa_projector}
\alias{print.summary.hatsa_projector}
\title{Print method for summary.hatsa_projector objects}
\usage{
\method{print}{summary.hatsa_projector}(x, ...)
}
\arguments{
\item{x}{A \code{summary.hatsa_projector} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for summary.hatsa_projector objects
}
