% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_hatsa_projector.R
\name{print.summary.task_hatsa_projector}
\alias{print.summary.task_hatsa_projector}
\title{Print method for summary.task_hatsa_projector objects}
\usage{
\method{print}{summary.task_hatsa_projector}(x, ...)
}
\arguments{
\item{x}{A `summary.task_hatsa_projector` object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for summary.task_hatsa_projector objects
}
