% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_quality_metrics.R
\name{print.summary_enhanced_hatsa}
\alias{print.summary_enhanced_hatsa}
\title{Print Enhanced Summary}
\usage{
\method{print}{summary_enhanced_hatsa}(x, ...)
}
\arguments{
\item{x}{Enhanced summary object}

\item{...}{Additional arguments}
}
\description{
Print Enhanced Summary
}
