% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_hatsa_projector.R
\name{print.task_hatsa_projector}
\alias{print.task_hatsa_projector}
\title{Print method for task_hatsa_projector objects}
\usage{
\method{print}{task_hatsa_projector}(x, ...)
}
\arguments{
\item{x}{A `task_hatsa_projector` object.}

\item{...}{Additional arguments passed to print.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for task_hatsa_projector objects
}
