% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{scores.hatsa_projector}
\alias{scores.hatsa_projector}
\title{Extract scores (stacked aligned sketches) from a hatsa_projector object}
\usage{
\method{scores}{hatsa_projector}(x, ...)
}
\arguments{
\item{x}{A \code{hatsa_projector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The stacked aligned sketches matrix (s).
}
\description{
Extract scores (stacked aligned sketches) from a hatsa_projector object
}
