% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_projector.R
\name{sdev.hatsa_projector}
\alias{sdev.hatsa_projector}
\title{Extract component standard deviations from a hatsa_projector object}
\usage{
sdev.hatsa_projector(x, ...)
}
\arguments{
\item{x}{A \code{hatsa_projector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A vector of component standard deviations (sdev).
}
\description{
Extract component standard deviations from a hatsa_projector object
}
