% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_simple.R
\name{select_anchors_auto}
\alias{select_anchors_auto}
\title{Automatic Anchor Selection Helper}
\usage{
select_anchors_auto(data, k, config)
}
\description{
Automatic Anchor Selection Helper
}
\keyword{internal}
