% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_hatsa_projector.R
\name{summary.task_hatsa_projector}
\alias{summary.task_hatsa_projector}
\title{Summary method for task_hatsa_projector objects}
\usage{
summary.task_hatsa_projector(object, ...)
}
\arguments{
\item{object}{A `task_hatsa_projector` object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list object of class `summary.task_hatsa_projector` containing summary statistics.
}
\description{
Summary method for task_hatsa_projector objects
}
