% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatsa_core_algorithm.R
\name{message_stage}
\alias{message_stage}
\title{Helper for printing stage messages, optionally only in interactive sessions}
\usage{
message_stage(message_text, verbose = TRUE, interactive_only = FALSE)
}
\arguments{
\item{interactive_only}{Logical, if TRUE, message only if session is interactive.}

\item{msg}{Message to print.}
}
\description{
Helper for printing stage messages, optionally only in interactive sessions
}
\keyword{internal}
