% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_stage}
\alias{message_stage}
\title{Helper for status messages}
\usage{
message_stage(message_text, verbose = TRUE, interactive_only = FALSE)
}
\arguments{
\item{message_text}{Text message to display.}
\item{verbose}{Whether to print the message.}
\item{interactive_only}{Print only in interactive mode.}
}
\description{
Prints a timestamped message when \code{verbose} is TRUE. If \code{interactive_only} is TRUE, the message is emitted only for interactive sessions.
}
\keyword{internal}
